finished KF Function in Puskesmas Role

+++ database/migrations/2025_11_12_074103_add_user_id_to_puskesmas_table.php
@@ -12,12 +12,14 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('puskesmas', function (Blueprint $table) {
-            $table->unsignedBigInteger('user_id')->after('id');
+            if (!Schema::hasColumn('puskesmas', 'user_id')) {
+                $table->unsignedBigInteger('user_id')->after('id');
 
-            $table->foreign('user_id')
-                ->references('id')
-                ->on('users')
-                ->onDelete('cascade');
+                $table->foreign('user_id')
+                    ->references('id')
+                    ->on('users')
+                    ->onDelete('cascade');
+            }
         });
     }
